General Structural Changes
Changed the structure of the html semantic content and stylesheet to
match. Changing everything to english for consistency.

Changed some identifiers to stick to clear and concise naming convention.

Added css folder for organization, changed name of imagenes folder to img
and scripts folder to js for consistency and proper naming conventions.

Added velocity js code though have not linked or implemented it yet.