Media Queries Fix
Fixed problem with screen size adaptations. Added multiple media queries
for a substantial range of screen size while maintaing a widescreen aspect
ratio and octal divisibilty (for clean resizing of graphics). If just this
resizing is enough that would be perfect... but somehow I predict that
I'll need to do a lot more work to make this responsive design flow
properly.