Added some background elements
Added background gradient for sky color, gradient for fade out from lands
color to background color and two divs taht surround the main wrapper div
to ensure that the game's background is seamless in all device sizes. Did
several calculations and cropping to ensure that the graphics measure
properly at different screen sizes. Will also have to do this for other
graphics and hopefully there won't be too much of a problem in generating
the canvas graphics with the right sizes.